% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTIR.R
\docType{data}
\name{FTIR}
\alias{FTIR}
\title{FTIR Spectra of A-, B-, and C-Type Starch}
\format{A data frame with 1038 rows and 4 variables}
\usage{
data(FTIR)
}
\description{
A dataset containing FTIR spectra of A-, B-, and C-type Starch.
}
\details{
\itemize{
  \item Wavelength. wavelength of FTIR spectrum
  \item A. Absorbance of an FTIR spectrum of A-type starch (A.U.).
  \item B. Absorbance of an FTIR spectrum of B-type starch (A.U.).
  \item C. Absorbance of an FTIR spectrum of C-type starch (A.U.).
}
}
\keyword{datasets}

