% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRD.R
\docType{data}
\name{XRD}
\alias{XRD}
\title{X-Ray Diffraction Patterns of A-, B-, and C-Type Starch}
\format{A data frame with 1527 rows and 4 variables}
\usage{
data(XRD)
}
\description{
A dataset containing X-ray diffraction patterns of A-, B-, and C-type Starch.
}
\details{
\itemize{
  \item Bragg_angle. bragg angle of X-ray diffraction patterns  (2\eqn{\theta}).
  \item A. intensity of an X-ray diffraction pattern of A-type starch (counts).
  \item B. intensity of an X-ray diffraction pattern of B-type starch (counts).
  \item C. intensity of an X-ray diffraction pattern of C-type starch (counts).
}
}
\keyword{datasets}

